refactor(utils): split llm_client into package
